requirements.txt
streamlit==1.3.0
pandas==1.5.1
requests==2.28.1
tensorflow==2.7.0
Pillow==9.3.0
scikit-learn==1.1.3
